requirements.txt
mcp[cli]
python-dotenv
ipykernel
httpx
starlette
requests